Refactor text copy component to be used outside of activation code context
